[Test] Split E2E nightly operator tests into RayCluster/GCS and RayJob runners (#3932)
