+++ ray-operator/test/e2eautoscaler/support.go
@@ -1,9 +1,6 @@
 package e2eautoscaler
 
 import (
-	"embed"
-
-	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	corev1ac "k8s.io/client-go/applyconfigurations/core/v1"
@@ -12,16 +9,6 @@ import (
 	rayv1ac "github.com/ray-project/kuberay/ray-operator/pkg/client/applyconfiguration/ray/v1"
 	. "github.com/ray-project/kuberay/ray-operator/test/support"
 )
-
-//go:embed *.py
-var _files embed.FS
-
-func ReadFile(t Test, fileName string) []byte {
-	t.T().Helper()
-	file, err := _files.ReadFile(fileName)
-	require.NoError(t.T(), err)
-	return file
-}
 
 type option[T any] func(t *T) *T
