+++ ray-operator/test/support/create_detached_placement_group.py
